+++ Examples/Java/src/main/java/com/aspose/barcode/cloud/client/examples/generatingSaving/cloudStorage/GenerateBarcodewithChecksumOption.java
@@ -1,4 +1,4 @@
-package com.aspose.barcode.cloud.client.examples.generating_saving.cloudStorage;
+package com.aspose.barcode.cloud.client.examples.generatingSaving.cloudStorage;
 
 import java.io.File;
 import java.io.IOException;
@@ -14,11 +14,11 @@ import com.aspose.barcode.model.SaaSposeResponse;
 import com.aspose.storage.api.StorageApi;
 
 /**
- * This sample code allows you to specify barcode image save format using
- * Aspose BarCode Cloud API.
- *   
+ * This sample code allows you to generate barcode with checksum using Aspose
+ * BarCode Cloud API.
+ * 
  */
-public class SpecifyBarcodeImageSaveFormat {
+public class GenerateBarcodewithChecksumOption {
 
 	public static void main(String[] args) {
 
@@ -31,7 +31,7 @@ public class SpecifyBarcodeImageSaveFormat {
 		// output folder
 		String outFolder = "c:\\temp\\";
 
-		InputStream inputStream = SpecifyBarcodeImageSaveFormat.class
+		InputStream inputStream = GenerateBarcodewithChecksumOption.class
 				.getClassLoader().getResourceAsStream(propFileName);
 		try {
 			if (inputStream != null) {
@@ -53,7 +53,6 @@ public class SpecifyBarcodeImageSaveFormat {
 		}
 
 		// ExStart:1
-
 		try {
 
 			// Instantiate Aspose.Storage Cloud API SDK
@@ -62,40 +61,40 @@ public class SpecifyBarcodeImageSaveFormat {
 			// Instantiate Aspose.BarCode Cloud API SDK
 			BarcodeApi barcodeApi = new BarcodeApi(apiKey, appSID, true);
 
-			// Set the barcode file name created on server
+			// Set the barcode file name created on server.
 			String name = "sample-barcode";
 			
-			// Set Text to encode inside barcode
+			// Set Text to encode inside barcode.
 			String text = "AsposeBarCode";
 			
 			// Set Barcode Symbology
 			String type = "Code128";
 			
 			// Set Generated Barcode Image Format
-			String format = "jpeg";
+			String format = "png";
 			
-			// Set Resolution along X and Y in dpi
-			Float resolutionX = 1.0f;
+			// Set Resolution along X and Y in dpi.
+			Float resolutionX = 0.0f;
 			Float resolutionY = 0.0f;
 			
-			// Set Width and Height of barcode unit
+			//Set Width and Height of barcode unit
 			Float dimensionX = 0.0f;
 			Float dimensionY = 0.0f;
 			
-			// Set Location, Measurement  of the code
-			String codeLocation = "";
+			//Set Location, Measurement  of the code
+			String codeLocation = "Above";
 			String grUnit = "mm";
 			
-			// Set if barcode's size will be updated automatically
+			// Sets if barcode's size will be updated automatically
 			String autoSize = "true";
 			
-			// Height of the bar.
+			// Set Height of the barcode
 			Float barHeight = 0.0f;
 			
-			// Set height, Width and quality of the image
+			//Set height, Width and quality of the image.
 			Float imageHeight = 0.0f;
 			Float imageWidth = 0.0f;
-			String imageQuality = "";
+			String imageQuality = "default";			
 			
 			// Set Angle of barcode orientation
 			Float rotAngle = 0.0f;
@@ -106,19 +105,19 @@ public class SpecifyBarcodeImageSaveFormat {
 			Float leftMargin = 0.0f;
 			Float rightMargin = 0.0f;
 			
-			// Set if checksum will be added to barcode image
-			String enableChecksum = "";
+			// Sets if checksum will be added to barcode image.
+			String enableChecksum = "Yes";
 			
 			//Set 3rd party cloud storage server (if any)
 			String storage = "";
-			
+
 			// Set folder location at cloud storage
 			String folder = "";
 			
 			File file = null;
 
-			// invoke Aspose.BarCode Cloud SDK API to generate image with
-			// specific barcode image format
+			// invoke Aspose.BarCode Cloud SDK API to generate barcode with
+			// checksum and save in cloud storage
 			SaaSposeResponse apiResponse = barcodeApi.PutBarcodeGenerateFile(
 					name, text, type, format, resolutionX, resolutionY,
 					dimensionX, dimensionY, codeLocation, grUnit, autoSize,
@@ -127,6 +126,7 @@ public class SpecifyBarcodeImageSaveFormat {
 					enableChecksum, storage, folder, file);
 
 			if (apiResponse != null && apiResponse.getStatus().equals("OK")) {
+			
 				// download generated barcode from cloud storage
 				com.aspose.storage.model.ResponseMessage storageRes = storageApi
 						.GetDownload(name, null, null);
