+++ Examples/Java/src/main/java/com/aspose/barcode/cloud/client/examples/generatingSaving/cloudStorage/SetBarcodeImageHeightWidthQualitySettings.java
@@ -1,4 +1,4 @@
-package com.aspose.barcode.cloud.client.examples.generating_saving.cloudStorage;
+package com.aspose.barcode.cloud.client.examples.generatingSaving.cloudStorage;
 
 import java.io.File;
 import java.io.IOException;
@@ -8,16 +8,17 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.util.Properties;
+
 import com.aspose.barcode.api.BarcodeApi;
 import com.aspose.barcode.model.SaaSposeResponse;
 import com.aspose.storage.api.StorageApi;
 
 /**
- * This sample code allows you to specify code text and symbology for barcode
- * according to specified parameters using Aspose BarCode Cloud API. 
+ * This sample code allows you to set barcode image height, width, and quality
+ * along with auto size option using Aspose BarCode Cloud API. 
  * 
  */
-public class SpecifyCodetextAndSymbologyForBarcode {
+public class SetBarcodeImageHeightWidthQualitySettings {
 
 	public static void main(String[] args) {
 
@@ -30,7 +31,7 @@ public class SpecifyCodetextAndSymbologyForBarcode {
 		// output folder
 		String outFolder = "c:\\temp\\";
 
-		InputStream inputStream = SpecifyCodetextAndSymbologyForBarcode.class
+		InputStream inputStream = SetBarcodeImageHeightWidthQualitySettings.class
 				.getClassLoader().getResourceAsStream(propFileName);
 		try {
 			if (inputStream != null) {
@@ -55,10 +56,10 @@ public class SpecifyCodetextAndSymbologyForBarcode {
 
 		try {
 
-			// Instantiate Aspose.Storage Cloud SDK
+			// Instantiate Aspose.Storage Cloud API SDK
 			StorageApi storageApi = new StorageApi(apiKey, appSID, true);
 
-			// Instantiate Aspose.BarCode Cloud SDK
+			// Instantiate Aspose.BarCode Cloud API SDK
 			BarcodeApi barcodeApi = new BarcodeApi(apiKey, appSID, true);
 
 			// Set the barcode file name created on server
@@ -68,33 +69,33 @@ public class SpecifyCodetextAndSymbologyForBarcode {
 			String text = "AsposeBarCode";
 			
 			// Set Barcode Symbology
-			String type = "Code128"; // symbology
+			String type = "Code128";
 			
 			// Set Generated Barcode Image Format
-			String format = "jpeg";
+			String format = "png";
 			
 			// Set Resolution along X and Y in dpi
-			Float resolutionX = 0.0f;
-			Float resolutionY = 0.0f;
+			Float resolutionX = 96.0f;
+			Float resolutionY = 96.0f;
 			
 			//Set Width and Height of barcode unit
-			Float dimensionX = 0.0f;
-			Float dimensionY = 0.0f;
+			Float dimensionX = 0.7f;
+			Float dimensionY = 2.0f;
 			
 			//Set Location, Measurement  of the code
-			String codeLocation = "";
+			String codeLocation = "Above";
 			String grUnit = "mm";
 			
 			// Sets if barcode's size will be updated automatically
 			String autoSize = "true";
 			
-			// Height of the bar
+			// Set Height of the bar.
 			Float barHeight = 0.0f;
 			
 			//Set height, Width and quality of the image
 			Float imageHeight = 0.0f;
 			Float imageWidth = 0.0f;
-			String imageQuality = "";
+			String imageQuality = "default";
 			
 			// Set Angle of barcode orientation
 			Float rotAngle = 0.0f;
@@ -106,7 +107,7 @@ public class SpecifyCodetextAndSymbologyForBarcode {
 			Float rightMargin = 0.0f;
 			
 			// Sets if checksum will be added to barcode image
-			String enableChecksum = "";
+			String enableChecksum = "Yes";
 			
 			//Set 3rd party cloud storage server (if any)
 			String storage = "";
@@ -114,10 +115,11 @@ public class SpecifyCodetextAndSymbologyForBarcode {
 			// Set folder location at cloud storage
 			String folder = "";
 			
+			// Set local file (if any)
 			File file = null;
 
 			// invoke Aspose.BarCode Cloud SDK API to generate image with
-			// specify code text and symbology for barcode
+			// specific height, width, and quality along with auto size option
 			SaaSposeResponse apiResponse = barcodeApi.PutBarcodeGenerateFile(
 					name, text, type, format, resolutionX, resolutionY,
 					dimensionX, dimensionY, codeLocation, grUnit, autoSize,
