+++ Examples/Java/src/main/java/com/aspose/barcode/cloud/client/examples/generatingSaving/cloudStorage/SetXAndYDimensionsOfBarcode.java
@@ -1,4 +1,4 @@
-package com.aspose.barcode.cloud.client.examples.generating_saving.cloudStorage;
+package com.aspose.barcode.cloud.client.examples.generatingSaving.cloudStorage;
 
 import java.io.File;
 import java.io.IOException;
@@ -14,11 +14,11 @@ import com.aspose.barcode.model.SaaSposeResponse;
 import com.aspose.storage.api.StorageApi;
 
 /**
- * This sample code allows you to set barcode image margins using Aspose BarCode
- * Cloud API.
+ * This sample code allows you to set X and Y dimensions of a barcode using
+ * Aspose BarCode Cloud API.
  * 
  */
-public class SetBarcodeImageMargin {
+public class SetXAndYDimensionsOfBarcode {
 
 	public static void main(String[] args) {
 
@@ -31,7 +31,7 @@ public class SetBarcodeImageMargin {
 		// output folder
 		String outFolder = "c:\\temp\\";
 
-		InputStream inputStream = SetBarcodeImageMargin.class
+		InputStream inputStream = SetXAndYDimensionsOfBarcode.class
 				.getClassLoader().getResourceAsStream(propFileName);
 		try {
 			if (inputStream != null) {
@@ -62,10 +62,10 @@ public class SetBarcodeImageMargin {
 			// Instantiate Aspose.BarCode Cloud SDK
 			BarcodeApi barcodeApi = new BarcodeApi(apiKey, appSID, true);
 
-			// Set the barcode file name created on server.
+			// Set the barcode file name created on server
 			String name = "sample-barcode";
 			
-			// Set Text to encode inside barcode.
+			// Set Text to encode inside barcode
 			String text = "AsposeBarCode";
 			
 			// Set Barcode Symbology
@@ -74,40 +74,40 @@ public class SetBarcodeImageMargin {
 			// Set Generated Barcode Image Format
 			String format = "png";
 			
-			// Set Resolution along X and Y in dpi.
-			Float resolutionX = 0.0f;
+			// Set Resolution along X and Y in dpi
+			Float resolutionX = 1.0f;
 			Float resolutionY = 0.0f;
 			
 			//Set Width and Height of barcode unit
-			Float dimensionX = 0.0f;
-			Float dimensionY = 0.0f;
+			Float dimensionX = 1.0f;
+			Float dimensionY = 0.5f;
 			
 			//Set Location, Measurement  of the code
-			String codeLocation = "Above";
+			String codeLocation = "";
 			String grUnit = "mm";
 			
 			// Sets if barcode's size will be updated automatically
 			String autoSize = "true";
-
+			
 			// Height of the bar.
 			Float barHeight = 0.0f;
 			
-			//Set height, Width and quality of the image.
+			//Set height, Width and quality of the image
 			Float imageHeight = 0.0f;
 			Float imageWidth = 0.0f;
-			String imageQuality = "default";
+			String imageQuality = "";
 			
 			// Set Angle of barcode orientation
 			Float rotAngle = 0.0f;
-			
+						
 			// Set Margin of image border
-			Float topMargin = 2.0f;
-			Float bottomMargin = 2.0f;
-			Float leftMargin = 2.0f;
-			Float rightMargin = 2.0f;
+			Float topMargin = 0.0f;
+			Float bottomMargin = 0.0f;
+			Float leftMargin = 0.0f;
+			Float rightMargin = 0.0f;
 			
-			// Sets if checksum will be added to barcode image.
-			String enableChecksum = "Yes";
+			// Sets if checksum will be added to barcode image
+			String enableChecksum = "";
 			
 			//Set 3rd party cloud storage server (if any)
 			String storage = "";
@@ -117,7 +117,8 @@ public class SetBarcodeImageMargin {
 			
 			File file = null;
 
-			// invoke Aspose.BarCode Cloud SDK API to generate image with specific image margins
+			// invoke Aspose.BarCode Cloud SDK API to generate image with
+			// specific X and Y dimensions of a barcode
 			SaaSposeResponse apiResponse = barcodeApi.PutBarcodeGenerateFile(
 					name, text, type, format, resolutionX, resolutionY,
 					dimensionX, dimensionY, codeLocation, grUnit, autoSize,
