+++ Examples/Java/src/main/java/com/aspose/barcode/cloud/client/examples/generatingSaving/withoutCloudStorage/GenerateBarcodeAndSaveImageToLocalDisk.java
@@ -1,4 +1,4 @@
-package com.aspose.barcode.cloud.client.examples.generating_saving.withoutCloudStorage;
+package com.aspose.barcode.cloud.client.examples.generatingSaving.withoutCloudStorage;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -12,11 +12,11 @@ import com.aspose.barcode.api.BarcodeApi;
 import com.aspose.barcode.model.ResponseMessage;
 
 /**
- * This sample code allows you to create barcode and save image to a stream
- * object using Aspose Cloud Storage. 
- *  
+ * This sample code allows you to create barcode and save image to local disk
+ * using Aspose BarCode Cloud API.
+ * 
  */
-public class GenerateBarcodeAndGetImageAsStream {
+public class GenerateBarcodeAndSaveImageToLocalDisk {
 
 	public static void main(String[] args) {
 
@@ -29,7 +29,7 @@ public class GenerateBarcodeAndGetImageAsStream {
 		// output folder
 		String outFolder = "c:\\temp\\";
 
-		InputStream inputStream = GenerateBarcodeAndGetImageAsStream.class
+		InputStream inputStream = GenerateBarcodeAndSaveImageToLocalDisk.class
 				.getClassLoader().getResourceAsStream(propFileName);
 		try {
 			if (inputStream != null) {
@@ -51,42 +51,43 @@ public class GenerateBarcodeAndGetImageAsStream {
 		}
 
 		// ExStart:1
-
 		try {
 
 			// Instantiate Aspose.BarCode Cloud API SDK
 			BarcodeApi barcodeApi = new BarcodeApi(apiKey, appSID, true);
 
-			// Set Filename of image
-			String name = "sample-barcode"; 
+			// Set Filename of image.
+			String name = "sample-barcode";
 			
 			// Set Text to encode inside barcode
-			String text = "Aspose.BarCode"; 
+			String text = "Aspose.BarCode";
 			
 			// Set Barcode Symbology
-			String type = "datamatrix"; 
+			String type = "QR";
 			
 			// Set Barcode Image Format
-			String format = "PNG"; 
-
-			// Sets if checksum will be added to barcode image.
-			String enableChecksum = "No"; 
-
-			//Set optional params (if any) 
-			Float resolutionX = null;
-			Float resolutionY = null;
-			Float dimensionX = null;
-			Float dimensionY = null;
+			String format = "PNG";
 			
+			// Set Resolution along X and Y in dpi.
+			Float resolutionX = 96.0f;
+			Float resolutionY = 96.0f;
+			
+			//Set Width and Height of barcode unit
+			Float dimensionX = 0.7f;
+			Float dimensionY = 2.0f;
+			
+			// Sets if checksum will be added to barcode image.
+			String enableChecksum = "NO";
 
-			// invoke Aspose.BarCode Cloud SDK API to create barcode and save image to a stream
+			// invoke Aspose.BarCode Cloud SDK API to create barcode and get
+			// barcode image as a stream
 			ResponseMessage apiResponse = barcodeApi.GetBarcodeGenerate(text,
 					type, format, resolutionX, resolutionY, dimensionX,
 					dimensionY, enableChecksum);
 
 			if (apiResponse != null && apiResponse.getStatus().equals("OK")) {
-				
-				// download generated barcode from api response
+
+				// download generated barcode image from api response
 				InputStream responseStream = apiResponse.getInputStream();
 				final Path destination = Paths.get(outFolder + name + "."
 						+ format);
