+++ Examples/Java/src/main/java/com/aspose/barcode/cloud/client/examples/managingRecognition/withoutCloudStorage/ReadBarcodeFromExternalImageURL.java
@@ -1,4 +1,4 @@
-package com.aspose.barcode.cloud.client.examples.managing_recognition.cloudStorage;
+package com.aspose.barcode.cloud.client.examples.managingRecognition.withoutCloudStorage;
 
 import java.io.File;
 import java.io.IOException;
@@ -8,13 +8,13 @@ import java.util.Properties;
 import com.aspose.barcode.api.BarcodeApi;
 import com.aspose.barcode.model.Barcode;
 import com.aspose.barcode.model.BarcodeResponseList;
-import com.aspose.storage.api.StorageApi;
 
 /**
- * This sample code allows you to read barcode using Aspose BarCode Cloud API.
+ * This sample code allows you to read barcode from external image URL using
+ * Aspose BarCode Cloud API.
  * 
  */
-public class ReadBarcodeFromAsposeCloudStorage {
+public class ReadBarcodeFromExternalImageURL {
 
 	public static void main(String[] args) {
 
@@ -27,7 +27,7 @@ public class ReadBarcodeFromAsposeCloudStorage {
 		// output folder
 		String outFolder = "c:\\temp\\";
 
-		InputStream inputStream = ReadBarcodeFromAsposeCloudStorage.class
+		InputStream inputStream = ReadBarcodeFromExternalImageURL.class
 				.getClassLoader().getResourceAsStream(propFileName);
 		try {
 			if (inputStream != null) {
@@ -52,54 +52,33 @@ public class ReadBarcodeFromAsposeCloudStorage {
 
 		try {
 
-			// Instantiate Aspose Storage Cloud API SDK
-			StorageApi storageApi = new StorageApi(apiKey, appSID, true);
-
 			// Instantiate Aspose BarCode Cloud API SDK
 			BarcodeApi barcodeApi = new BarcodeApi(apiKey, appSID, true);
 
-			// set input file name
-			String name = "sample-barcode.jpeg";
-			
 			//The barcode type.
 			//If this parameter is empty, autodetection of all supported types is used.
 			String type = "";
 			
-			//Sets mode for checksum validation during recognition
+			//Set mode for checksum validation during recognition
 			String checksumValidation = "";
-			
-			//Sets if FNC symbol stripping should be performed. 
-			Boolean stripFnc = Boolean.TRUE;
 
+			//Sets if FNC symbol stripping should be performed
+			Boolean stripFnc = false;
+			
 			//Sets recognition of rotated barcode
-			Integer rotationAngle = Integer.MIN_VALUE;
+			Integer rotationAngle = 0;
 			
-			//Sets exact number of barcodes to recognize 
-			Integer barcodesCount = Integer.MIN_VALUE;
+			//Set the image file url 
+			String url = "http://www.barcoding.com/images/Barcodes/code93.gif";
 			
-			//Set recognition of barcode inside specified Rectangle region
-			Integer rectX = Integer.SIZE;
-			Integer rectY = Integer.SIZE;			
-			Integer rectWidth = Integer.MIN_VALUE;
-			Integer rectHeight = Integer.MIN_VALUE;
-			
-			//Set 3rd party cloud storage server (if any)
-			String storage = "";
-			
-			// Set folder location at cloud storage
-			String folder = "";
+			File file = null;
 
-			// upload files to aspose cloud storage
-			storageApi.PutCreate(name, "", "",
-					new File(ReadBarcodeFromAsposeCloudStorage.class
-							.getResource("/" + name).toURI()));
-
-			// invoke Aspose.BarCode Cloud SDK API to read barcode from Aspose
-			// Cloud Storage
-			BarcodeResponseList apiResponse = barcodeApi.GetBarcodeRecognize(
-					name, type, checksumValidation, stripFnc, rotationAngle,
-					barcodesCount, rectX, rectY, rectWidth, rectHeight,
-					storage, folder);
+			// invoke Aspose.BarCode Cloud SDK API to read barcode from external
+			// image URL
+			BarcodeResponseList apiResponse = barcodeApi
+					.PostBarcodeRecognizeFromUrlorContent(type,
+							checksumValidation, stripFnc, rotationAngle, url,
+							file);
 
 			if (apiResponse != null && apiResponse.getStatus().equals("OK")) {
