+++ Examples/Java/src/main/java/com/aspose/barcode/cloud/client/examples/managingRecognition/withoutCloudStorage/ReadBarcodeFromLocalFile.java
@@ -1,25 +1,21 @@
-package com.aspose.barcode.cloud.client.examples.managing_recognition.withoutCloudStorage;
+package com.aspose.barcode.cloud.client.examples.managingRecognition.withoutCloudStorage;
 
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
 
 import com.aspose.barcode.api.BarcodeApi;
-import com.aspose.barcode.cloud.client.examples.managing_recognition.cloudStorage.ReadBarcodeFromAsposeCloudStorage;
 import com.aspose.barcode.model.Barcode;
 import com.aspose.barcode.model.BarcodeReader;
 import com.aspose.barcode.model.BarcodeResponseList;
-import com.aspose.barcode.model.BinarizationHints;
 import com.aspose.barcode.model.ChecksumValidation;
-import com.aspose.storage.api.StorageApi;
 
 /**
- * This sample code allows users to read barcode by apply various available
- * image processing algorithms using Aspose BarCode Cloud API.
+ * This sample code allows you to read barcode from local image using Aspose BarCode Cloud API.
  * 
+
  */
-public class ReadBarcodesbyApplyingImageProcessingAlgorithm {
+public class ReadBarcodeFromLocalFile {
 
 	public static void main(String[] args) {
 
@@ -32,7 +28,7 @@ public class ReadBarcodesbyApplyingImageProcessingAlgorithm {
 		// output folder
 		String outFolder = "c:\\temp\\";
 
-		InputStream inputStream = ReadBarcodeFromAsposeCloudStorage.class
+		InputStream inputStream = ReadBarcodeFromLocalFile.class
 				.getClassLoader().getResourceAsStream(propFileName);
 		try {
 			if (inputStream != null) {
@@ -57,40 +53,29 @@ public class ReadBarcodesbyApplyingImageProcessingAlgorithm {
 
 		try {
 
-			// Instantiate Aspose Storage Cloud API SDK
-			StorageApi storageApi = new StorageApi(apiKey, appSID, true);
-
 			// Instantiate Aspose BarCode Cloud API SDK
 			BarcodeApi barcodeApi = new BarcodeApi(apiKey, appSID, true);
 
-			// Set the barcode file name created on server
+			// set input file name
 			String name = "sample-barcode.jpeg";
 
 			// The barcode type.
-			// If this parameter is empty, autodetection of all supported types is used
+			// If this parameter is empty, autodetection of all supported types
+			// is used.
 			String type = "";
 
-			// Set folder location at cloud storage
 			String folder = "";
 
 			BarcodeReader body = new BarcodeReader();
 
-			// Set if FNC symbol stripping should be performed.
+			// Sets if FNC symbol stripping should be performed
 			body.setStripFNC(true);
 
 			// Set mode for checksum validation during recognition
 			body.setChecksumValidation(ChecksumValidation.OFF);
 
-			// Set special mode of barcode binarization
-			body.setBinarizationHints(BinarizationHints.ComplexBackground);
-
-			// upload files to aspose cloud storage
-			storageApi.PutCreate(name, "", "", new File(
-					ReadBarcodesbyApplyingImageProcessingAlgorithm.class
-							.getResource("/" + name).toURI()));
-
 			// invoke Aspose.BarCode Cloud SDK API to recognition of a barcode
-			// by apply various available image processing algorithms
+			// from file on server with parameters in body
 			BarcodeResponseList apiResponse = barcodeApi
 					.PutBarcodeRecognizeFromBody(name, type, folder, body);
