+++ Examples/Java/src/main/java/com/aspose/barcode/cloud/client/examples/managingRecognition/withoutCloudStorage/ReadBarcodesWithChecksum.java
@@ -1,4 +1,4 @@
-package com.aspose.barcode.cloud.client.examples.managing_recognition.withoutCloudStorage;
+package com.aspose.barcode.cloud.client.examples.managingRecognition.withoutCloudStorage;
 
 import java.io.File;
 import java.io.IOException;
@@ -6,15 +6,22 @@ import java.io.InputStream;
 import java.util.Properties;
 
 import com.aspose.barcode.api.BarcodeApi;
+import com.aspose.barcode.cloud.client.examples.managingRecognition.cloudStorage.ReadBarcodeFromAsposeCloudStorage;
 import com.aspose.barcode.model.Barcode;
+import com.aspose.barcode.model.BarcodeReader;
 import com.aspose.barcode.model.BarcodeResponseList;
+import com.aspose.barcode.model.BinarizationHints;
+import com.aspose.barcode.model.ChecksumValidation;
+import com.aspose.storage.api.StorageApi;
 
 /**
- * This sample code allows you to read barcode from external image URL using
- * Aspose BarCode Cloud API.
+ * This sample code allows you to read barcode using Aspose BarCode Cloud API.
+ * You just need to use Aspose.Cells for Cloud API in any language of your
+ * choice.
  * 
+ * @author Farooq Sheikh
  */
-public class ReadBarcodeFromExternalImageURL {
+public class ReadBarcodesWithChecksum {
 
 	public static void main(String[] args) {
 
@@ -27,7 +34,7 @@ public class ReadBarcodeFromExternalImageURL {
 		// output folder
 		String outFolder = "c:\\temp\\";
 
-		InputStream inputStream = ReadBarcodeFromExternalImageURL.class
+		InputStream inputStream = ReadBarcodeFromAsposeCloudStorage.class
 				.getClassLoader().getResourceAsStream(propFileName);
 		try {
 			if (inputStream != null) {
@@ -52,33 +59,41 @@ public class ReadBarcodeFromExternalImageURL {
 
 		try {
 
+			// Instantiate Aspose Storage Cloud API SDK
+			StorageApi storageApi = new StorageApi(apiKey, appSID, true);
+
 			// Instantiate Aspose BarCode Cloud API SDK
 			BarcodeApi barcodeApi = new BarcodeApi(apiKey, appSID, true);
 
+			// set input file name
+			String name = "sample-barcode.jpeg";
+			
 			//The barcode type.
 			//If this parameter is empty, autodetection of all supported types is used.
 			String type = "";
 			
-			//Set mode for checksum validation during recognition
-			String checksumValidation = "";
+			String folder = "";
 
-			//Sets if FNC symbol stripping should be performed
-			Boolean stripFnc = false;
-			
-			//Sets recognition of rotated barcode
-			Integer rotationAngle = 0;
-			
-			//Set the image file url 
-			String url = "http://www.barcoding.com/images/Barcodes/code93.gif";
-			
-			File file = null;
+			BarcodeReader body = new BarcodeReader();
 
-			// invoke Aspose.BarCode Cloud SDK API to read barcode from external
-			// image URL
+			// Set if FNC symbol stripping should be performed. 
+			body.setStripFNC(true);
+			
+			// Set mode for checksum validation during recognition
+			body.setChecksumValidation(ChecksumValidation.ON);
+			
+			// Set special mode of barcode binarization
+			body.setBinarizationHints(BinarizationHints.ComplexBackground);
+
+			// upload files to aspose cloud storage
+			storageApi.PutCreate(name, "", "", new File(
+					ReadBarcodesWithChecksum.class.getResource("/" + name)
+							.toURI()));
+
+			// invoke Aspose.BarCode Cloud SDK API to recognition of a barcode
+			// from file on server with parameters in body
 			BarcodeResponseList apiResponse = barcodeApi
-					.PostBarcodeRecognizeFromUrlorContent(type,
-							checksumValidation, stripFnc, rotationAngle, url,
-							file);
+					.PutBarcodeRecognizeFromBody(name, type, folder, body);
 
 			if (apiResponse != null && apiResponse.getStatus().equals("OK")) {
