+++ Examples/Java/src/main/java/com/aspose/barcode/cloud/client/examples/managingRecognition/withoutCloudStorage/ReadBarcodesbyApplyingImageProcessingAlgorithm.java
@@ -1,21 +1,25 @@
-package com.aspose.barcode.cloud.client.examples.managing_recognition.withoutCloudStorage;
+package com.aspose.barcode.cloud.client.examples.managingRecognition.withoutCloudStorage;
 
+import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
 
 import com.aspose.barcode.api.BarcodeApi;
+import com.aspose.barcode.cloud.client.examples.managingRecognition.cloudStorage.ReadBarcodeFromAsposeCloudStorage;
 import com.aspose.barcode.model.Barcode;
 import com.aspose.barcode.model.BarcodeReader;
 import com.aspose.barcode.model.BarcodeResponseList;
+import com.aspose.barcode.model.BinarizationHints;
 import com.aspose.barcode.model.ChecksumValidation;
+import com.aspose.storage.api.StorageApi;
 
 /**
- * This sample code allows you to read barcode from local image using Aspose BarCode Cloud API.
+ * This sample code allows users to read barcode by apply various available
+ * image processing algorithms using Aspose BarCode Cloud API.
  * 
-
  */
-public class ReadBarcodeFromLocalFile {
+public class ReadBarcodesbyApplyingImageProcessingAlgorithm {
 
 	public static void main(String[] args) {
 
@@ -28,7 +32,7 @@ public class ReadBarcodeFromLocalFile {
 		// output folder
 		String outFolder = "c:\\temp\\";
 
-		InputStream inputStream = ReadBarcodeFromLocalFile.class
+		InputStream inputStream = ReadBarcodeFromAsposeCloudStorage.class
 				.getClassLoader().getResourceAsStream(propFileName);
 		try {
 			if (inputStream != null) {
@@ -53,29 +57,40 @@ public class ReadBarcodeFromLocalFile {
 
 		try {
 
+			// Instantiate Aspose Storage Cloud API SDK
+			StorageApi storageApi = new StorageApi(apiKey, appSID, true);
+
 			// Instantiate Aspose BarCode Cloud API SDK
 			BarcodeApi barcodeApi = new BarcodeApi(apiKey, appSID, true);
 
-			// set input file name
+			// Set the barcode file name created on server
 			String name = "sample-barcode.jpeg";
 
 			// The barcode type.
-			// If this parameter is empty, autodetection of all supported types
-			// is used.
+			// If this parameter is empty, autodetection of all supported types is used
 			String type = "";
 
+			// Set folder location at cloud storage
 			String folder = "";
 
 			BarcodeReader body = new BarcodeReader();
 
-			// Sets if FNC symbol stripping should be performed
+			// Set if FNC symbol stripping should be performed.
 			body.setStripFNC(true);
 
 			// Set mode for checksum validation during recognition
 			body.setChecksumValidation(ChecksumValidation.OFF);
 
+			// Set special mode of barcode binarization
+			body.setBinarizationHints(BinarizationHints.ComplexBackground);
+
+			// upload files to aspose cloud storage
+			storageApi.PutCreate(name, "", "", new File(
+					ReadBarcodesbyApplyingImageProcessingAlgorithm.class
+							.getResource("/" + name).toURI()));
+
 			// invoke Aspose.BarCode Cloud SDK API to recognition of a barcode
-			// from file on server with parameters in body
+			// by apply various available image processing algorithms
 			BarcodeResponseList apiResponse = barcodeApi
 					.PutBarcodeRecognizeFromBody(name, type, folder, body);
